refactor: migrate to new toast system
