+++ app/src/composables/errorHandler.ts
@@ -1,13 +1,13 @@
-import { useToastStore } from '@/stores/toast'
+import { useToastStore } from '@/stores/useToastStore'
 import { ToastType } from '@/types'
 
 export function useErrorHandler() {
-  const { show } = useToastStore()
+  const { addToast } = useToastStore()
 
   function handleError(error: any) {
     console.log('Error:', error.message)
     error.value = error.message
-    show(ToastType.Error, error.message)
+    addToast({ type: ToastType.Error, message: error.message, timeout: 5000 })
   }
 
   return { handleError }
